requirement.txt
streamlit
youtube-transcript-api

langchain
langchain-core
langchain-community
langchain-text-splitters
langchain-huggingface

faiss-cpu
sentence-transformers
transformers

huggingface-hub

python-dotenv
